🐛 Fix: 데이터 패칭 제거
